Rename `get_trait_info` to `self_clause_for_item`
